pyproject.toml
[project]
name = "examen-bentoml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bentoml>=1.4.28",
    "fastapi>=0.121.1",
    "pandas>=2.3.3",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.38.0",
]
